+++ R/data_tabulate.R
@@ -159,6 +159,14 @@ data_tabulate.default <- function(
   # save label attribute, before it gets lost...
   var_label <- attr(x, "label", exact = TRUE)
 
+  # save by attribute
+  if (!is.null(by)) {
+    by_name <- tryCatch(
+      insight::safe_deparse(substitute(by)),
+      error = function(e) NULL
+    )
+  }
+
   # save and fix variable name, check for grouping variable
   obj_name <- tryCatch(
     insight::safe_deparse(substitute(x)),
@@ -183,11 +191,10 @@ data_tabulate.default <- function(
 
   # we go into another function for crosstables here...
   if (!is.null(by)) {
-    by_name <- tryCatch(
-      insight::safe_deparse(substitute(by)),
-      error = function(e) NULL
-    )
+    # don't lose that name of the by variable
+    attr(x, "by") <- by_name
     by <- .validate_by(by, x)
+
     return(.crosstable(
       x,
       by = by,
@@ -195,11 +202,10 @@ data_tabulate.default <- function(
       remove_na = remove_na,
       proportions = proportions,
       obj_name = obj_name,
-      by_name = by_name,
       group_variable = group_variable
-    ))
+    )
+    )
   }
-
   # frequency table
   if (is.null(weights)) {
     if (remove_na) {
@@ -293,6 +299,11 @@ data_tabulate.default <- function(
 
   attr(out, "total_n") <- sum(out$N, na.rm = TRUE)
   attr(out, "valid_n") <- valid_n
+  if (!is.null(by)) {
+    attr(out, "by") <- by_name
+  } else {
+    attr(out, "by") <- NULL
+  }
 
   class(out) <- c("datawizard_table", "data.frame")
 
@@ -317,6 +328,13 @@ data_tabulate.data.frame <- function(
   verbose = TRUE,
   ...
 ) {
+  if (!is.null(by)) {
+    by_name <- tryCatch(
+      insight::safe_deparse(substitute(by)),
+      error = function(e) NULL
+    )
+    by_name <- gsub('\\"', "", by_name)
+  }
   # evaluate arguments
   select <- .select_nse(
     select,
@@ -327,10 +345,10 @@ data_tabulate.data.frame <- function(
     verbose = verbose
   )
 
-  by_name <- tryCatch(
-    insight::safe_deparse(substitute(by)),
-    error = function(e) NULL
-  )
+  if (!is.null(by)) {
+    attr(x, "by") <- by_name
+  }
+
   # validate "by"
   by <- .validate_by(by, x)
   # validate "weights"
@@ -349,13 +367,13 @@ data_tabulate.data.frame <- function(
       ...
     )
   })
-
   if (is.null(by)) {
     class(out) <- c("datawizard_tables", "list")
   } else {
+    out <- lapply(out, structure, by = by_name)
     class(out) <- c("datawizard_crosstabs", "list")
-    attr(out, "by") <- gsub('\\"', "", by_name, fixed = TRUE)
   }
+
   attr(out, "collapse") <- isTRUE(collapse)
   attr(out, "is_weighted") <- !is.null(weights)
 
@@ -382,7 +400,14 @@ data_tabulate.grouped_df <- function(
   grps <- attr(x, "groups", exact = TRUE)
   group_variables <- data_remove(grps, ".rows")
   grps <- grps[[".rows"]]
-
+  # save the by variable name
+  if (!is.null(by)) {
+    by_name <- tryCatch(
+      insight::safe_deparse(substitute(by)),
+      error = function(e) NULL
+    )
+    by_name <- gsub('\\"', "", by_name)
+  }
   # evaluate arguments
   select <- .select_nse(
     select,
@@ -391,11 +416,6 @@ data_tabulate.grouped_df <- function(
     ignore_case,
     regex = regex,
     verbose = verbose
-  )
-
-  by_name <- tryCatch(
-    insight::safe_deparse(substitute(by)),
-    error = function(e) NULL
   )
 
   x <- as.data.frame(x)
@@ -430,8 +450,8 @@ data_tabulate.grouped_df <- function(
   if (is.null(by)) {
     class(out) <- c("datawizard_tables", "list")
   } else {
+    out <- lapply(out, structure, by = by_name)
     class(out) <- c("datawizard_crosstabs", "list")
-    attr(out, "by") <- gsub('\\"', "", by_name, fixed = TRUE)
   }
   attr(out, "collapse") <- isTRUE(collapse)
   attr(out, "is_weighted") <- !is.null(weights)