+++ R/data_xtabulate.R
@@ -7,12 +7,16 @@
   remove_na = FALSE,
   proportions = NULL,
   obj_name = NULL,
-  by_name = NULL,
   group_variable = NULL
 ) {
+
   if (!is.null(proportions)) {
     proportions <- match.arg(proportions, c("row", "column", "full"))
   }
+  if (!is.null(attr(x, "by"))) {
+    by_name <- attr(x, "by")
+  }
+
   # frequency table
   if (is.null(weights)) {
     # we have a `.default` and a `.data.frame` method for `data_tabulate()`.
@@ -90,7 +94,10 @@
   attr(out, "weights") <- weights
   attr(out, "proportions") <- proportions
   attr(out, "varname") <- obj_name
-  attr(out, "by") <- by_name
+
+  if (!is.null(by_name)) {
+    attr(out, "by") <- by_name
+  }
 
   attr(out, "grouped_df") <- !is.null(group_variable)
   attr(out, "prop_table") <- .prop_table(out)